add soft deletes for job offers

+++ app/Models/JobOffer.php
@@ -4,11 +4,12 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 use Illuminate\Support\Arr;
 
 class JobOffer extends Model
 {
-    use HasFactory;
+    use HasFactory, SoftDeletes;
 
     protected $guarded = [];
 
@@ -45,7 +46,10 @@ class JobOffer extends Model
 
         self::deleting(function(JobOffer $jobOffer)
         {
-            $jobOffer->tags()->detach();
+            if ($jobOffer->isForceDeleting())
+            {
+                $jobOffer->tags()->detach();
+            }
         });
     }
